+++ src/main/java/ru/vyarus/gradle/frontend/core/model/root/JsResource.java
@@ -35,5 +35,10 @@ public class JsResource extends RootResource {
     protected ResourceMinifier getMinifier() {
         return new JsMinifier();
     }
+
+    @Override
+    protected String getFileExtension() {
+        return "js";
+    }
 }