+++ CHANGELOG.md
@@ -1,2 +1,5 @@
+* Fix downloading urls without extension declared
+  (for example, google font https://fonts.googleapis.com/css?family=Roboto)
+
 ### 1.0.0 (2023-09-28)
 * Initial release